+++ src/types.ts
@@ -167,6 +167,14 @@ export interface DatasetQuery {
     "template-tags"?: Record<string, unknown>;
   };
   query?: Record<string, unknown>;
+  stages?: (Record<string, unknown> & {
+    native?:
+      | string
+      | {
+          query: string;
+          "template-tags"?: Record<string, unknown>;
+        };
+  })[];
 }
 
 export interface DatasetResponse {